+++ src/components/consts.js
@@ -1,0 +1,8 @@
+const TASK_COUNT = 3;
+
+const Place = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+export {TASK_COUNT, Place};
